- deleted numeric type of sequencetracksegment, we now only use sequencetracksegmentcurve types for float, vec2, vec3 and vec4 types

+++ apps/timeline/module/src/sequencetrack.h
@@ -15,7 +15,7 @@ namespace nap
 	enum SequenceTrackTypes : int
 	{
 		UNKOWN,
-		NUMERIC,
+		FLOAT,
 		VEC2,
 		VEC3,
 		VEC4
